fix: address sidepanel review feedback for panel selection and graph reuse
Agent-Logs-Url: https://github.com/Qredence/fleet-rlm/sessions/9d62fe5e-ceb9-4d88-9438-9d627c72dac9
