+++ src/frontend/src/features/workspace/inspection/message-inspector-panel.tsx
@@ -7,13 +7,13 @@ import {
   buildAssistantContentModel,
   type AssistantContentModel,
 } from "@/features/workspace/conversation/assistant-content/model";
-import type { ExecutionStep } from "@/features/workspace/use-workspace";
 import { useChatStore } from "@/features/workspace/use-workspace";
 import { useWorkspaceUiStore } from "@/features/workspace/use-workspace";
 import type { InspectorTab } from "@/features/workspace/use-workspace";
 import { executionSectionState } from "./inspector-ui";
 
 import { ExecutionInspectorTab } from "./tabs/execution-inspector-tab";
+import { hasMeaningfulGraph } from "./tabs/graph-inspector-content";
 import { GraphInspectorTab } from "./tabs/graph-inspector-tab";
 import { MessageInspectorTab } from "./tabs/message-inspector-tab";
 
@@ -37,29 +37,6 @@ function EmptyInspectorState({ hasAssistantTurns }: { hasAssistantTurns: boolean
       </Card>
     </div>
   );
-}
-
-function hasMeaningfulGraph(steps: ExecutionStep[]) {
-  if (steps.length < 2) return false;
-
-  const lanes = new Set(
-    steps
-      .map((step) => step.lane_key ?? `${step.actor_kind ?? "unknown"}:${step.actor_id ?? ""}`)
-      .filter(Boolean),
-  );
-  if (lanes.size > 1) return true;
-
-  if (steps.some((step) => step.actor_kind === "delegate" || step.actor_kind === "sub_agent")) {
-    return true;
-  }
-
-  const childCounts = new Map<string, number>();
-  for (const step of steps) {
-    if (!step.parent_id) continue;
-    childCounts.set(step.parent_id, (childCounts.get(step.parent_id) ?? 0) + 1);
-  }
-
-  return [...childCounts.values()].some((count) => count > 1);
 }
 
 function selectedTurnStatus(
